interpreter: add instruction loop
The runner cannot be halted in the middle of the execution.

This commit adds an interface to Runner that can be used by function.coffee
to later pause and resume the execution of the program at will.

This is the first of a series of commits to implement a step-by-step
runner that can be halted at the instruction level.